requirements.txt
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1 
psycopg2==2.8.4
python-decouple==3.3 
drf-yasg
pandas
sqlalchemy
boto3
